refactor: 💡 rename table to modules
